# Restore RawResponse speed